backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
diskcache==5.6.3
Django==6.0.3
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.17.0
drf-spectacular==0.29.0
et_xmlfile==2.0.0
gunicorn==25.3.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
psycopg2==2.9.11
PyMuPDF==1.27.2.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
xlrd==2.0.2
